# shop/api_specs/api_spec.json
{
    "swagger": "2.0",
    "host": "127.0.0.1:8000",
    "basePath": "/api/shop/",
    "info": {
        "version": "1.0.0",
        "title": "Simple API",
        "description": "A simple API to learn how to write OpenAPI Specification"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "security": [],
    "securityDefinitions": {
      "oauth": {
          "tokenUrl": "http://auth.ibtspl.com/oauth2/",
          "flow": "password",
          "scopes": {
              "read": "read users",
              "write": "create users",
              "update": "update users",
              "delete": "delete users",
              "superuser": "super user permission" 
          },
          "type": "oauth2"
      }
  },
    "definitions": {
...
        "ProductId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "BasicProduct": {
            "type": "object",
            "properties": {
                "name":{
                  "type":"string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format":"float"
                },
                "mfg_date": {
                    "type": "string"
                },
                "exp_date": {
                    "type": "string"
                },
                "category":{
                     "type":"string"
                },
                "stock_quantity": {
                    "type": "number",
                    "format":"int32"
                }
            },
            "required": [
                "description",
                "name",
                "stock_quantity",
                "exp_date",
                "mfg_date",
                "price"
            
            ]
        },
        "Product": {
            "allOf": [
                {
                    "$ref": "#/definitions/BasicProduct"
                },
                {
                    "$ref": "#/definitions/ProductId"
                }
            ]
        },
...
        "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" }
        }
      },
...
      "CustomerWithExtraFields": {
        "allOf": [
            {
                "$ref": "#/definitions/Customer"
            },
            {
                "type": "object",
                "properties": {
                    "createdOn": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "ProductId": {
                        "type": "string"
                    }
                },
                "required": [
                    "createdOn",
                    "ProductId"
                ]
            }
        ]
    },
      "Payment": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "order_id": { "type": "string" },
          "amount": { "type": "number", "format": "float"},
          "method": { "type": "string", "enum": ["Card", "NetBanking", "UPI", "COD"] },
          "status": { "type": "string", "enum": ["Success", "Pending", "Failed"] },
          "transaction_date": { "type": "string", "format": "date-time" }
        }
      },
...
"CreatePaymentRequest": {
  "type": "object",
  "properties": {
    "amount": {
      "type": "number",
      "format": "float",
      "description": "The payment amount for the order."
    },
    "method": {
      "type": "string",
      "enum": ["Credit Card", "Debit Card", "PayPal", "Bank Transfer"],
      "description": "The method of payment."
    }
  },
  "required": [ "amount", "method"]
},
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "products": {
            "type": "array",
            "items": { "$ref": "#/definitions/Product" }
          },
          "customer": { "$ref": "#/definitions/Customer" },
          "payment": { "$ref": "#/definitions/Payment" },
          "status": { "type": "string", "enum": ["Pending", "Shipped", "Delivered", "Cancelled"] },
          "expected_delivery_date": { "type": "string", "format": "date" },
          "order_datetime": { "type": "string", "format": "date-time" }
        }
      },
...
    },
 "parameters": {
...
        "CreateNewCustomerParameter": {
            "name": "customer",
            "in": "body",
            "description": "The customer to create.",
            "schema": {
                "$ref": "#/definitions/Customer"
            }
        },
        "CreateNewOrderParameter": {
          "name":"order",
    "in": "body",
    "description": "The order details including the selected products and their quantities.",
    "required": true,
    "schema": {
      "$ref": "#/definitions/Order"
    }
  },
  "CreateNewPaymentParameter": {
    "name":"payment",
    "in": "body",
    "description": "The payment details including the order ID, amount, and payment method.",
    "required": true,
    "schema": {
      "$ref": "#/definitions/CreatePaymentRequest"
    }
  },
...
    },
    "responses": {
...
        "CreateNewCustomerResponse": {
            "description": "Success Response",
            "schema": {
                "$ref": "#/definitions/CustomerWithExtraFields"
            }
        },
        "CreateNewOrderResponse": {
    "description": "Order created successfully.",
    "schema": {
      "$ref": "#/definitions/Order"
    }
  },
  "CreateNewPaymentResponse": {
    "description": "Payment recorded successfully.",
    "schema": {
      "$ref": "#/definitions/Payment"
    }
  },
...
    },
    "paths": {
...
          "/product/delete/{id}/": {
      "delete": {
        "summary": "Delete a product",
        "description": "Deletes a product by its ID.",
      "operationId": "delete_product",
      "security": [
                    {
                        "oauth": ["delete"]
                    }
                ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the product to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          }
        }
      }
           },
...
                "/customer/create/": {
                  "post": {
                      "operationId": "create_customer",
                      "summary": "Create New customer",
                      "description": "Creates a new product and return the customer object. This api is not idempotent.",
                      "security": [
                    {
                        "oauth": ["write"]
                    }
                ],
                      "parameters": [
                          {
                              "$ref": "#/parameters/CreateNewCustomerParameter"
                          }
                      ],
                      "responses": {
                          "200": {
                              "$ref": "#/responses/CreateNewCustomerResponse"
                          }
                      }
                  }
              },
...
               "/customers/delete/{id}/":{
                "delete": {
              "operationId": "delete_customer",
              "summary": "Delete a customer",
              "description": "Remove a customer from the system.",
              "security": [
                    {
                        "oauth": ["delete"]
                    }
                ],
              "parameters": [
                { "name": "id", "in": "path", "required": true,"type": "string" }
              ],
              "responses": {
                "204": { "description": "Customer successfully deleted." }
              }
            }
               },
...
"/order/create/": {
  "post": {
    "operationId": "create_order",
    "summary": "Create New Order",
    "description": "Creates a new order and returns the order object. This API is not idempotent.",
    "security": [
                    {
                        "oauth": ["write"]
                    }
                ],
    "parameters": [
      {
        "$ref": "#/parameters/CreateNewOrderParameter"
      }
    ],
    "responses": {
      "200": {
        "$ref": "#/responses/CreateNewOrderResponse"
      }
    }
  }
},
...
"/orders/{orderId}/cancel/": {
  "put": {
    "summary": "Cancel an order",
    "description": "Mark an order as canceled and update inventory if necessary.",
    "operationId": "cancel_order",
    "security": [
                    {
                        "oauth": ["update"]
                    }
                ],
    "parameters": [
      {
        "name": "orderId",
        "in": "path",
        "description": "The ID of the order to cancel.",
        "required": true,
        "type": "integer"
      },
      {
        "name": "reason",
        "in": "body",
        "description": "Optional reason for order cancellation.",
        "required": false,
        "schema": {
          "type": "object",
          "properties": {
            "reason": {
              "type": "string",
              "example": "Customer request."
            }
          }
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Order successfully canceled.",
        "schema": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Order canceled successfully."
            },
            "orderId": {
              "type": "integer",
              "example": 123
            },
            "status": {
              "type": "string",
              "example": "Canceled"
            },
            "inventoryUpdated": {
              "type": "boolean",
              "example": true
            }
          }
        }
      },
      "400": {
        "description": "Invalid request or order cannot be canceled.",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "example": "Order is already shipped or delivered and cannot be canceled."
            }
          }
        }
      },
      "404": {
        "description": "Order not found.",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "example": "Order not found."
            }
          }
        }
      }
    }
  }
},
...
"/payment/create/": {
  "post": {
    "operationId": "create_payment",
    "summary": "Record New Payment",
    "description": "Adds a payment record for a specific order and returns the payment object. This API is not idempotent.",
    "security": [
                    {
                        "oauth": ["write"]
                    }
                ],
    "parameters": [
      {
        "$ref": "#/parameters/CreateNewPaymentParameter"
      }
    ],
    "responses": {
      "200": {
        "$ref": "#/responses/CreateNewPaymentResponse"
      },
      "400": {
        "description": "Invalid input data.",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "example": "Invalid payment amount provided."
            }
          }
        }
      }
    }
  }
},
...
"/feedback/delete/{feedbackId}/": {
  "delete": {
    "summary": "Delete feedback",
    "description": "Remove a customer's feedback from the system.",
    "operationId": "delete_feedback",
    "security": [
                    {
                        "oauth": ["delete"]
                    }
                ],
    "parameters": [
      {
        "name": "feedbackId",
        "in": "path",
        "description": "The ID of the feedback to delete.",
        "required": true,
        "type": "integer"
      }
    ],
    "responses": {
      "200": {
        "description": "Feedback deleted successfully.",
        "schema": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Feedback deleted successfully."
            }
          }
        }
      },
      "404": {
        "description": "Feedback not found.",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "example": "Feedback with ID 101 not found."
            }
          }
        }
      }
    }
  }
}
}
}