# shop/models/payment.py
from django.db import models

class Payment(models.Model):
    amount = models.DecimalField(max_digits=10, decimal_places=2)
    method = models.CharField(max_length=50)
    transaction_date=models.DateField(null=True, blank=True)