# shop/views/delete_feedback/api_wrapper.py
from dsu.dsu_gen.openapi.decorator.interface_decorator import \
    validate_decorator
from .validator_class import ValidatorClass
from shop.models import Feedback
from django.http import JsonResponse


@validate_decorator(validator_class=ValidatorClass)
def api_wrapper(*args, **kwargs):
    # ---------MOCK IMPLEMENTATION---------
    feedback_id=kwargs['path_params']['id']
    try:
        feedback=Feedback.objects.get(id=feedback_id)
        feedback.delete()

        return JsonResponse({'Feedback is deleted...'},status=200)
    
    except Feedback.DoesNotExist:
        return JsonResponse({'error': 'Feedback not found'}, status=404)


    try:
        from shop.views.delete_feedback.request_response_mocks \
            import REQUEST_BODY_JSON
        body = REQUEST_BODY_JSON
    except ImportError:
        body = {}

    test_case = {
        "path_params": {'feedbackId': 873},
        "query_params": {},
        "header_params": {},
        "body": body,
        "securities": []
    }

    from dsu.dsu_gen.openapi.utils.mock_response import mock_response

    try:
        response = ''
        status_code = 200
        if '200' in ['200', '404']:
            from shop.views.delete_feedback.request_response_mocks \
                import RESPONSE_200_JSON
            response = RESPONSE_200_JSON
            status_code = 200
        elif '201' in ['200', '404']:
            from shop.views.delete_feedback.request_response_mocks \
                import RESPONSE_201_JSON
            response = RESPONSE_201_JSON
            status_code = 201
    except ImportError:
        response = ''
        status_code = 200
    response_tuple = mock_response(
        app_name="shop", test_case=test_case,
        operation_name="delete_feedback",
        kwargs=kwargs, default_response_body=response,
        group_name="", status_code=status_code)
    return response_tuple